refactor: DTO 경로 변경, stationService 합치고 메서드 축소
