+++ src/main/kotlin/subway/service/StationService.kt
@@ -1,18 +1,21 @@
 package subway.service
 
+import subway.domain.Station
 import subway.repository.SectionRepository
 import subway.repository.StationRepository
 
-object DeleteStationService {
-    private fun validate(name: String) {
+object StationService {
+
+    fun register(name: String) {
+        require(!StationRepository.existsByName(name))
+        StationRepository.add(Station(name))
+    }
+
+    fun delete(name: String) {
         require(
             SectionRepository.findAll()
                 .none { it.matchUpward(name) || it.matchDownward(name) }
         )
-    }
-
-    fun delete(name: String) {
-        validate(name)
         StationRepository.deleteByName(name)
     }
 }
